+++ pkgs/ffigen/lib/src/header_parser/sub_parsers/functiondecl_parser.dart
@@ -21,7 +21,8 @@ List<Func>? parseFunctionDeclaration(clang_types.CXCursor cursor) {
 
   final funcUsr = cursor.usr();
   final funcName = cursor.spelling();
-  if (shouldIncludeFunc(funcUsr, funcName)) {
+  final decl = Declaration(usr: funcUsr, originalName: funcName);
+  if (shouldIncludeFunc(decl)) {
     _logger.fine('++++ Adding Function: ${cursor.completeStringRepr()}');
 
     final returnType = cursor.returnType().toCodeGenType();
@@ -49,8 +50,7 @@ List<Func>? parseFunctionDeclaration(clang_types.CXCursor cursor) {
       parameters.add(
         Parameter(
           originalName: paramName,
-          name:
-              config.functionDecl.renameMemberUsingConfig(funcName, paramName),
+          name: config.functionDecl.renameMember(decl, paramName),
           type: paramType,
         ),
       );
@@ -110,17 +110,16 @@ List<Func>? parseFunctionDeclaration(clang_types.CXCursor cursor) {
           nesting.length + commentPrefix.length,
         ),
         usr: funcUsr + vaFunc.postfix,
-        name: config.functionDecl.renameUsingConfig(funcName) + vaFunc.postfix,
+        name: config.functionDecl.rename(decl) + vaFunc.postfix,
         originalName: funcName,
         returnType: returnType,
         parameters: parameters,
         varArgParameters:
             vaFunc.types.map((ta) => Parameter(type: ta, name: 'va')).toList(),
         exposeSymbolAddress:
-            config.functionDecl.shouldIncludeSymbolAddress(funcName),
-        exposeFunctionTypedefs:
-            config.exposeFunctionTypedefs.shouldInclude(funcName),
-        isLeaf: config.leafFunctions.shouldInclude(funcName),
+            config.functionDecl.shouldIncludeSymbolAddress(decl),
+        exposeFunctionTypedefs: config.shouldExposeFunctionTypedef(decl),
+        isLeaf: config.isLeafFunction(decl),
         objCReturnsRetained: objCReturnsRetained,
         ffiNativeConfig: config.ffiNativeConfig,
       ));