+++ pkgs/ffigen/test/config_tests/exclude_all_by_default_test.dart
@@ -2,8 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:ffigen/ffigen.dart';
 import 'package:ffigen/src/code_generator.dart';
+import 'package:ffigen/src/header_parser.dart' show parse;
 import 'package:ffigen/src/strings.dart' as strings;
 import 'package:test/test.dart';