+++ pkgs/ffigen/test/native_objc_test/block_test.dart
@@ -24,7 +24,6 @@ import 'util.dart';
 void main() {
   group('Blocks', () {
     setUpAll(() {
-      logWarnings();
       // TODO(https://github.com/dart-lang/native/issues/1068): Remove this.
       DynamicLibrary.open('../objective_c/test/objective_c.dylib');
       final dylib = File('test/native_objc_test/block_test.dylib');