+++ pkgs/ffigen/test/native_objc_test/failed_to_load_test.dart
@@ -13,7 +13,6 @@ import 'util.dart';
 void main() {
   group('Failed to load', () {
     setUpAll(() {
-      logWarnings();
       generateBindingsForCoverage('failed_to_load');
     });