+++ pkgs/ffigen/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 14.0.0-wip
+
+- Create a public facing API for ffigen that can be invoked as a library:
+  `void generate(Config config)`. Make `Config` an implementatble interface,
+  rather than needing to be parsed from yaml.
+
 ## 13.0.0
 
 - __Breaking change__: Code-gen the ObjC `id` type to `ObjCObjectBase` rather